+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v2.5.2
+    * Fixed deploy script for Foundation-security-services-setup. Distributed to all repos.
+    * Added ARCHITECTURE.md, describing the operation of the OpenSecOps Installer and its scripts.
+    * Added header doc strings to all scripts under scripts/.
+
 ## v2.5.1
     * Fixed script execution to properly pass --verbose flag to component scripts when using verbose mode.